Create PUMA
PUMA Token  

"PUMA token is a community-based digital asset and is not affiliated with, endorsed by, or related to PUMA SE, the global sportswear company. The use of the name is purely symbolic and represents community strength and agility." 

PUMA Token is a decentralized digital asset built on the Binance Smart Chain (BEP-20). Designed for speed, security, and scalability, PUMA enables seamless peer-to-peer transactions and supports the development of decentralized applications in Web3 ecosystems. Its vision is to empower digital finance, gaming, e-commerce, and online communities through transparency, innovation, and strong community engagement. With a fair token distribution, liquidity provisioning, and future deflationary mechanisms such as token burns, PUMA is designed to grow sustainably and reward its holders.